Refactor instructor-audit-logs-page.component.spec
